\name{igvR}
\alias{igvR}
\alias{igvR-package}
\title{igvR:  Control IGV (the Integrated Genomics Viewer) from R}
\description{
  IGV, from the Broad Institute, can be controlled though a
  TCP/IP socket port, using simple commands.  This package
  provides an R implementation of this remote control.
  }

 \keyword{Dynamic Graphics}

